# Avoid allocation with the offset option for `String#unpack1`