@maximecb 2024-05-02
Thanks Alan. Very cool :D

We might need to regenerate the benchmark in yjit-bench as well.